feat: add domain status
