+++ packages/base-types/src/models/version/domain.ts
@@ -1,0 +1,14 @@
+export enum DomainStatus {
+  DESIGN = 'DESIGN',
+  REVIEW = 'REVIEW',
+  COMPLETE = 'COMPLETE',
+}
+
+export interface Domain {
+  id: string;
+  live: boolean;
+  name: string;
+  status?: DomainStatus;
+  topicIDs: string[];
+  rootDiagramID: string;
+}
